requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.4
django-anymail==5.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0